Don't rebuild countries.json in the rebuilder
This doesn't work as expected because there isn't a commit after the
data is rebuilt so there's nothing to update in countries.json.

This also had another subtle problem where if there was nothing to be
rebuilt for the country but there was a change to countries.json it
would make a commit just updating the countries.json file, which isn't
the job of this repo.