Only create Pull Request if Popolo has changed
If there are no changes to ep-popolo-v1.0.json then don't create the
Pull Request on everypolitician-data.

This takes the simple route of allowing the branch to be created first,
and then comparing the branch with master to see what has changed. This
means that there will be some branches that get pushed but don't get a
Pull Request created for them because they don't change the Popolo.
Perhaps this is a good thing since it means that if you want/need to
create a Pull Request for a non-Popolo change you can do so manually.