Make the output from the build hang around for longer
We've only got a single worker doing the rebuilds and the pull requests,
so when it gets busy it can be more than an hour between the rebuild
happening and the pull request being created, which was leading to the
output from the build getting lost.

This temporarily puts that timeout up to 6 hours while we figure out
better ways to speed up the pipeline.