Add rebuild helper function
This adds a shortcut for creating a rebuild for a specific country.
Having this in a helper method gives us the flexibility to add extra
checks in the future without duplicating that logic in multiple routes.

The method returns a string so that it can be used as the last line in a
route, meaning the string will become the response body.