Move build steps into a shell script
This moves the bulk of the logic that deals with interacting with the
everypolitician-data repo into a shell script. This is because we were
basically trying to emulate a shell script with the ExternalCommand
class, but we were doing a poor job because we weren't always checking
for errors etc. By farming this job out to bash we can reduce the size
of the interface between the rebuilder and the everypolitician-data
repo.

This should make it much easier to:

- Test the full rebuilder locally without having to clone the full
  ep-data repo
- Change the way the rebuilder works so we're not performing a full
  clone and bundle install on each request
- Get much better test coverage