Do a shallow clone of everypolitician-data repo
A large proportion of the build time is currently spent cloning the
everypolitician-data repo, which is now at least 1GB in size on the
wire. To reduce the amount of unnecessary data that is being moved
around this change switches to a shallow clone of the repo.

Because we're using a shallow clone I've also had to disable the
countries.json build step, because that relies on git history to work
currently. We'll need to either remove the dependency on git history, or
have it happen at a different stage.