Log which source we're working with
Rather than only providing info on the source if we hit an error, log it
immediately we begin processing. This will make it easier to see what's
currently in progress if tailing the live logs, or to see which
sources are taking how much time.

Before this, we might get something like:

```
2018-09-26T09:46:17.416568+00:00 app[worker.1]: 4 TID-gox59p9u4 RebuilderJob JID-d20612a57cf57749530ea5d5 INFO: start
2018-09-26T09:47:53.263840+00:00 app[worker.1]: 4 TID-gox59p9u4 RebuilderJob JID-d20612a57cf57749530ea5d5 INFO: done: 95.847 sec
```

which tells us not very much at all.